Stabilized The Display
Adjusted the code so it generates a constant 262 scanline frame by adding timers and moving code so it run during the VBLANK and overscan periods instead of before or afterwards.

The previous code was generating 254 to 255 scanlines per frame, depending upon how may CPU cycles were used between the overscan and VBLANK periods.